admin reg

+++ app/controllers/workersController.js
@@ -3,14 +3,14 @@
 
     angular.module('app').controller('workersController', Controller);
 
-    function Controller($window, crudService,$scope, $http, $uibModal,  FlashService) {
+    function Controller($window, crudService,$scope, $uibModal,  FlashService) {
         console.log("workersController loaded");
         var workCtrl = this;
 
         workCtrl.worker = null;
 
         workCtrl.saveUser = saveUser;
-        workCtrl.deleteUser = deleteUser;
+        // workCtrl.deleteUser = deleteUser;
 
         initController();
 
@@ -30,6 +30,7 @@
         };
 
 // Modal windows
+
         this.animationsEnabled = true;
         this.openAdd = function () {
             var modalInstance = $uibModal.open({
@@ -72,16 +73,7 @@
             });
         };
 
-        // Crud worker
-        $scope.removeWorker = function (id) {
-            crudService.removeWorker(id);
-            crudService.getWorkers($scope.currentPage, $scope.itemsPerPage).success(function (response) {
-                $scope.totalItems = response.total;
-                $scope.currentPage = response.page;
-                $scope.workers = response.docs;
-                $scope.itemsPerPage = response.limit;
-            });
-        };
+        // CRUD
         $scope.pageChanged = function () {
             crudService.getWorkers($scope.currentPage, $scope.itemsPerPage).success(function (response) {
                 $scope.totalItems = response.total;
@@ -98,29 +90,53 @@
                 $scope.itemsPerPage = response.limit;
                 $scope.currentPage = 1; //reset to first page
             });
-        }
-
-        // Crud user
+        };
         function saveUser() {
             crudService.Update(workCtrl.worker).then(function () {
-                    FlashService.Success('User updated');
-                })
+                FlashService.Success('User updated');
+            })
                 .catch(function (error) {
                     FlashService.Error(error);
                 });
 
         };
 
-        function deleteUser() {
-            crudService.Delete(workCtrl.worker._id)
+            $scope.removeWorker = function (_id) {
+                // log admin out
+                crudService.GetCurrent().then(function (user) {
+                    workCtrl.worker = user;
+                    console.log(workCtrl.worker._id, _id);
+                    if(workCtrl.worker._id === _id)
+                    {
+                        $window.location = '/login';
+                    }
+                });
+            crudService.Delete(_id)
                 .then(function () {
-                    // log user out
-                    $window.location = '/login';
                 })
                 .catch(function (error) {
                     FlashService.Error(error);
                 });
+            // console.log(workCtrl.worker.role);
+            if(workCtrl.worker.role==='Пользователь' || workCtrl.worker.role===undefined)
+            {
+                crudService.Delete(workCtrl.worker._id)
+                    .then(function () {
+                    })
+                    .catch(function (error) {
+                        FlashService.Error(error);
+                    });
+                // log user out
+                $window.location = '/login';
+            }
+                crudService.getWorkers($scope.currentPage, $scope.itemsPerPage).success(function (response) {
+                    $scope.totalItems = response.total;
+                    $scope.currentPage = response.page;
+                    $scope.workers = response.docs;
+                    $scope.itemsPerPage = response.limit;
+                });
         };
+
 
     };
